backend/requirement.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
python-dotenv
deepface
opencv-python
numpy
email-validator
google-genai